+++ Bencodex.Tests/Types/DictionaryTest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Text;
 using Bencodex.Types;
 using Xunit;
 
@@ -288,6 +289,44 @@ namespace Bencodex.Tests.Types
                 () => dictionary.GetValue<Binary>(listKey));
             Assert.Throws<InvalidCastException>(
                 () => dictionary.GetValue<Text>(listKey));
+        }
+
+        [Fact]
+        public void Inspection()
+        {
+            Assert.Equal("{}", Dictionary.Empty.Inspection);
+
+            var one = Dictionary.Empty.SetItem("foo", "bar");
+            Assert.Equal(
+                "{\n  \"foo\": \"bar\"\n}",
+                one.Inspection
+            );
+            Assert.Equal(
+                "{\n  b\"\\x66\\x6f\\x6f\": \"bar\"\n}",
+                Dictionary.Empty.SetItem(Encoding.ASCII.GetBytes("foo"), "bar").Inspection
+            );
+            Assert.Equal(
+                @"{
+  ""baz"": {
+    ""foo"": ""bar""
+  },
+  ""foo"": ""bar""
+}",
+                one.SetItem("baz", one).Inspection
+            );
+        }
+
+        [Fact]
+        public void String()
+        {
+            Assert.Equal(
+                "Bencodex.Types.Dictionary {}",
+                Dictionary.Empty.ToString()
+            );
+            Assert.Equal(
+                "Bencodex.Types.Dictionary {\n  \"foo\": \"bar\"\n}",
+                Dictionary.Empty.SetItem("foo", "bar").ToString()
+            );
         }
     }
 }